Move `Rope` implementation to it's own file
